requirements.txt
flask
qrcode
python-dotenv